Move dtype related things to its own file
Also implement the conversion of ExSeriesDtype to Polars' Datatype.
