+++ native/explorer/src/datatypes.rs
@@ -1,3 +1,5 @@
+mod ex_dtypes;
+
 use crate::atoms;
 use crate::ExplorerError;
 use chrono::prelude::*;
@@ -15,6 +17,8 @@ use std::str::FromStr;
 
 #[cfg(feature = "aws")]
 use polars::prelude::cloud::AmazonS3ConfigKey as S3Key;
+
+pub use ex_dtypes::*;
 
 pub struct ExDataFrameRef(pub DataFrame);
 pub struct ExExprRef(pub Expr);
@@ -422,95 +426,6 @@ impl TryFrom<ExParquetCompression> for ParquetCompression {
 
         Ok(compression)
     }
-}
-
-impl rustler::Encoder for Box<ExSeriesDtype> {
-    fn encode<'a>(&self, env: rustler::Env<'a>) -> rustler::Term<'a> {
-        let dtype: &ExSeriesDtype = self.deref();
-
-        dtype.encode(env)
-    }
-}
-
-impl<'a> rustler::Decoder<'a> for Box<ExSeriesDtype> {
-    fn decode(term: rustler::Term<'a>) -> rustler::NifResult<Self> {
-        let dtype: ExSeriesDtype = term.decode()?;
-        Ok(Box::new(dtype))
-    }
-}
-
-#[derive(NifTaggedEnum)]
-pub enum ExTimeUnit {
-    Millisecond,
-    Microsecond,
-    Nanosecond,
-}
-
-#[derive(NifTaggedEnum)]
-pub enum ExSeriesDtype {
-    Binary,
-    Boolean,
-    Category,
-    Date,
-    Float,
-    Integer,
-    String,
-    Time,
-    Datetime(ExTimeUnit),
-    Duration(ExTimeUnit),
-    List(Box<ExSeriesDtype>),
-}
-
-impl TryFrom<&DataType> for ExSeriesDtype {
-    type Error = ExplorerError;
-
-    fn try_from(value: &DataType) -> Result<Self, Self::Error> {
-        match value {
-            DataType::Binary => Ok(ExSeriesDtype::Binary),
-            DataType::Boolean => Ok(ExSeriesDtype::Boolean),
-            DataType::Categorical(_) => Ok(ExSeriesDtype::Category),
-            DataType::Date => Ok(ExSeriesDtype::Date),
-            DataType::Float64 => Ok(ExSeriesDtype::Float),
-            DataType::Int64 => Ok(ExSeriesDtype::Integer),
-            DataType::Time => Ok(ExSeriesDtype::Time),
-            DataType::Utf8 => Ok(ExSeriesDtype::String),
-            DataType::Datetime(TimeUnit::Nanoseconds, _) => {
-                Ok(ExSeriesDtype::Datetime(ExTimeUnit::Nanosecond))
-            }
-            DataType::Datetime(TimeUnit::Microseconds, _) => {
-                Ok(ExSeriesDtype::Datetime(ExTimeUnit::Microsecond))
-            }
-            DataType::Datetime(TimeUnit::Milliseconds, _) => {
-                Ok(ExSeriesDtype::Datetime(ExTimeUnit::Millisecond))
-            }
-
-            DataType::Duration(TimeUnit::Nanoseconds) => {
-                Ok(ExSeriesDtype::Duration(ExTimeUnit::Nanosecond))
-            }
-            DataType::Duration(TimeUnit::Microseconds) => {
-                Ok(ExSeriesDtype::Duration(ExTimeUnit::Microsecond))
-            }
-            DataType::Duration(TimeUnit::Milliseconds) => {
-                Ok(ExSeriesDtype::Duration(ExTimeUnit::Millisecond))
-            }
-
-            DataType::List(inner) => Ok(ExSeriesDtype::List(Box::new(Self::try_from(
-                inner.as_ref(),
-            )?))),
-
-            _ => Err(ExplorerError::Other(format!(
-                "cannot cast to dtype: {value}"
-            ))),
-        }
-    }
-}
-
-// Unsigned, Signed, Float
-#[derive(NifTaggedEnum)]
-pub enum ExSeriesIoType {
-    U(u8),
-    S(u8),
-    F(u8),
 }
 
 // =========================
